+++ Nursery/NUBranchAperture.m
@@ -56,7 +56,7 @@
 
 - (NUUInt64)nextFixedOOP
 {
-    NUIvar *anIvar = [[self character] ivarInAllOOPIvarsAt:currentFixedOOPIvarIndex++];
+    NUIvar *anIvar = [[self character] ivarInAllOOPIvarsAt:(NSUInteger)currentFixedOOPIvarIndex++];
     return [pupilNote readUInt64At:[anIvar offset]];
 }
